refactor(config): address review feedback on dogstatsdAddr migration
- Move resolveDogstatsdAddr from ddtrace/tracer/option.go into
  internal/config/config_helpers.go. The pure helper takes the
  agent-reported statsd port as an int (instead of agentFeatures) so
  it has no dependency on the tracer package. Constants for the default
  statsd port and the default agent hostname now live in internal/config
  and internal/agent respectively.
- Add Config.ResolveDogstatsdAddr method that resolves, stores, and
  reports telemetry (origin = OriginCalculated). The c.statsdClient != nil
  precedence is preserved at the option.go call site: resolution only
  runs when no custom statsd client was provided.
- Switch the setter telemetry key from DD_DOGSTATSD_HOST to
  DD_DOGSTATSD_URL, since the resolved value is a URL/host:port rather
  than just a hostname. Register DD_DOGSTATSD_URL in
  internal/env/supported_configurations.json and regenerate the gen.go.
- Trim the dogstatsdAddr field comment and remove the redundant getter
  doc comment per reviewer feedback.
- Move TestResolveDogstatsdAddr to internal/config/config_helpers_test.go
  alongside the function it tests; adapt its table to the new signature.

Addresses review comments on #4741.
