+++ ddtrace/tracer/option_test.go
@@ -1023,149 +1023,6 @@ func TestDefaultHTTPClient(t *testing.T) {
 	})
 }
 
-func TestResolveDogstatsdAddr(t *testing.T) {
-	socketFile, err := os.CreateTemp("", "dsd.socket")
-	require.NoError(t, err)
-	require.NoError(t, socketFile.Close())
-	t.Cleanup(func() { os.RemoveAll(socketFile.Name()) })
-	socketPath := socketFile.Name()
-
-	tests := []struct {
-		name       string
-		configAddr string
-		af         agentFeatures
-		env        map[string]string
-		socketPath string
-		expected   string
-	}{
-		{
-			name:     "defaults",
-			expected: "localhost:8125",
-		},
-		{
-			name:     "host-env",
-			env:      map[string]string{"DD_DOGSTATSD_HOST": "111.111.1.1", "DD_AGENT_HOST": "222.222.2.2"},
-			expected: "111.111.1.1:8125",
-		},
-		{
-			name:     "port-env",
-			env:      map[string]string{"DD_DOGSTATSD_PORT": "8111"},
-			expected: "localhost:8111",
-		},
-		{
-			name:     "port-env+agent-host-env",
-			env:      map[string]string{"DD_DOGSTATSD_PORT": "8111", "DD_AGENT_HOST": "222.222.2.2"},
-			expected: "222.222.2.2:8111",
-		},
-		{
-			name:     "host-env+port-env",
-			env:      map[string]string{"DD_DOGSTATSD_HOST": "111.111.1.1", "DD_DOGSTATSD_PORT": "8888", "DD_AGENT_HOST": "222.222.2.2"},
-			expected: "111.111.1.1:8888",
-		},
-		{
-			name:       "host-env+socket",
-			env:        map[string]string{"DD_DOGSTATSD_HOST": "111.111.1.1"},
-			socketPath: socketPath,
-			expected:   "111.111.1.1:8125",
-		},
-		{
-			name:       "port-env+socket",
-			env:        map[string]string{"DD_DOGSTATSD_PORT": "8111"},
-			socketPath: socketPath,
-			expected:   "localhost:8111",
-		},
-		{
-			name:       "socket",
-			socketPath: socketPath,
-			expected:   "unix://" + socketPath,
-		},
-		// DD_AGENT_HOST alone should not trigger the env var path;
-		// it falls through to auto-discovery.
-		{
-			name:       "agent-host-env-only+socket",
-			env:        map[string]string{"DD_AGENT_HOST": "222.222.2.2"},
-			socketPath: socketPath,
-			expected:   "unix://" + socketPath,
-		},
-		{
-			name:     "agent-host-env-only",
-			env:      map[string]string{"DD_AGENT_HOST": "222.222.2.2"},
-			expected: "222.222.2.2:8125",
-		},
-		{
-			name:     "agent-host-env-only+agent-port",
-			env:      map[string]string{"DD_AGENT_HOST": "222.222.2.2"},
-			af:       agentFeatures{StatsdPort: 9876},
-			expected: "222.222.2.2:9876",
-		},
-		// configAddr (priority 1) wins over everything.
-		{
-			name:       "config-addr",
-			configAddr: "custom:9999",
-			expected:   "custom:9999",
-		},
-		{
-			name:       "config-addr+env",
-			configAddr: "custom:9999",
-			env:        map[string]string{"DD_DOGSTATSD_HOST": "111.111.1.1", "DD_DOGSTATSD_PORT": "8111"},
-			expected:   "custom:9999",
-		},
-		{
-			name:       "config-addr+socket",
-			configAddr: "custom:9999",
-			socketPath: socketPath,
-			expected:   "custom:9999",
-		},
-		{
-			name:       "config-addr+agent-port",
-			configAddr: "custom:9999",
-			af:         agentFeatures{StatsdPort: 9876},
-			expected:   "custom:9999",
-		},
-		// Agent-reported port used as fallback when env host is set but no env port.
-		{
-			name:     "host-env+agent-port",
-			env:      map[string]string{"DD_DOGSTATSD_HOST": "111.111.1.1"},
-			af:       agentFeatures{StatsdPort: 9876},
-			expected: "111.111.1.1:9876",
-		},
-		// Env port wins over agent-reported port.
-		{
-			name:     "host-env+port-env+agent-port",
-			env:      map[string]string{"DD_DOGSTATSD_HOST": "111.111.1.1", "DD_DOGSTATSD_PORT": "8111"},
-			af:       agentFeatures{StatsdPort: 9876},
-			expected: "111.111.1.1:8111",
-		},
-		// Auto-discovery: agent-reported port when no env and no socket.
-		{
-			name:     "agent-port",
-			af:       agentFeatures{StatsdPort: 9876},
-			expected: "localhost:9876",
-		},
-		// Auto-discovery: socket wins over agent-reported port.
-		{
-			name:       "socket+agent-port",
-			af:         agentFeatures{StatsdPort: 9876},
-			socketPath: socketPath,
-			expected:   "unix://" + socketPath,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Clear all relevant env vars so that each subtest is isolated
-			// from the host environment and other subtests.
-			for _, key := range []string{"DD_DOGSTATSD_HOST", "DD_DOGSTATSD_PORT", "DD_AGENT_HOST"} {
-				t.Setenv(key, "")
-			}
-			for k, v := range tt.env {
-				t.Setenv(k, v)
-			}
-			assert.Equal(t, tt.expected, resolveDogstatsdAddr(tt.configAddr, tt.af, tt.socketPath))
-		})
-	}
-}
-
 func TestServiceName(t *testing.T) {
 	t.Run("WithService", func(t *testing.T) {
 		defer globalconfig.SetServiceName("")
